requirements.txt
python-dotenv
requests
flask
werkzeug
gunicorn